+++ LenovoLegionToolkit.Lib/Station/Logging/IExtensionLogger.cs
@@ -1,0 +1,9 @@
+using System;
+
+namespace LenovoLegionToolkit.Lib.Station.Logging;
+
+public interface IExtensionLogger
+{
+    void Trace(string message);
+    void Error(string message, Exception exception);
+}
