+++ LenovoLegionToolkit.Lib/Station/Services/IUiDispatcher.cs
@@ -1,0 +1,10 @@
+using System;
+using System.Threading.Tasks;
+
+namespace LenovoLegionToolkit.Lib.Station.Services;
+
+public interface IUiDispatcher
+{
+    Task InvokeAsync(Action action);
+    Task<T> InvokeAsync<T>(Func<T> action);
+}
