--- LenovoLegionToolkit.Lib/Utils/IExtensionProvider.cs
@@ -1,12 +1,0 @@
-using System.Threading.Tasks;
-
-namespace LenovoLegionToolkit.Lib.Utils;
-
-public interface IExtensionProvider
-{
-    void Initialize(object context);
-    Task ExecuteAsync(string action, params object[] args);
-    object? GetData(string key);
-    void SetData(string key, object value);
-    void Dispose();
-}
