+++ LenovoLegionToolkit.WPF/App.xaml.cs
@@ -33,7 +33,10 @@ using LenovoLegionToolkit.Lib.Settings;
 using LenovoLegionToolkit.Lib.SoftwareDisabler;
 using LenovoLegionToolkit.Lib.System;
 using LenovoLegionToolkit.Lib.Utils;
+using LenovoLegionToolkit.Lib.Station.Core;
 using LenovoLegionToolkit.WPF.CLI;
+using LenovoLegionToolkit.WPF.Station.Core;
+using LenovoLegionToolkit.WPF.Station.Services;
 using LenovoLegionToolkit.WPF.Extensions;
 using LenovoLegionToolkit.WPF.Resources;
 using LenovoLegionToolkit.WPF.Utils;
