+++ LenovoLegionToolkit.WPF/IoCModule.cs
@@ -4,6 +4,9 @@ using LenovoLegionToolkit.Lib.Utils;
 using LenovoLegionToolkit.WPF.CLI;
 using LenovoLegionToolkit.WPF.Settings;
 using LenovoLegionToolkit.WPF.Utils;
+using LenovoLegionToolkit.Lib.Station.Core;
+using LenovoLegionToolkit.Lib.Station.Logging;
+using LenovoLegionToolkit.Lib.Station.Services;
 
 namespace LenovoLegionToolkit.WPF;
 
@@ -24,10 +27,10 @@ public class IoCModule : Module
 
         builder.Register<IpcServer>();
 
-        builder.RegisterType<Extensions.NavigationService>().As<INavigationService>().SingleInstance();
-        builder.RegisterType<Extensions.ExtensionManager>().SingleInstance();
-        builder.RegisterType<Extensions.ExtensionContextFactory>().SingleInstance();
-        builder.RegisterType<Extensions.ExtensionLogger>().As<IExtensionLogger>();
-        builder.RegisterType<Extensions.UiDispatcher>().As<IUiDispatcher>().SingleInstance();
+        builder.RegisterType<Station.Services.NavigationService>().As<INavigationService>().SingleInstance();
+        builder.RegisterType<Station.Core.ExtensionManager>().SingleInstance();
+        builder.RegisterType<Station.Core.ExtensionContextFactory>().SingleInstance();
+        builder.RegisterType<Station.Logging.ExtensionLogger>().As<IExtensionLogger>();
+        builder.RegisterType<Station.Services.UiDispatcher>().As<IUiDispatcher>().SingleInstance();
     }
 }
